* fix postinst scripts for RHEL

+++ base/conf/yaffas-certificates/redhat/yaffas-certificates.spec
@@ -24,38 +24,40 @@ make install DESTDIR=$RPM_BUILD_ROOT
 %{__rm} -rf $RPM_BUILD_ROOT
 
 %post
-set -e
-CERTDIR="/opt/yaffas/etc/ssl/certs/"
-CERTS="postfix exim webmin cyrus usermin ldap zarafa-webaccess zarafa-server zarafa-gateway zarafa-ical mppserver"
-for i in $CERTS; do
-	ln -sf $CERTDIR/org/default.key $CERTDIR/$i.key
-	ln -sf $CERTDIR/org/default.crt $CERTDIR/$i.crt
-	cat $CERTDIR/org/default.crt $CERTDIR/org/default.key > $CERTDIR/$i
-done
-PRODUCTS="webmin usermin"
-for PRODUCT in $PRODUCTS; do
-	CONF="/opt/yaffas/etc/$PRODUCT/miniserv.conf"
-	TMPCONF="/tmp/miniserv.conf"
-	if [ -r $CONF ]; then
-		echo "Editing $PRODUCT"
-		KC="key cert"
-		for BBKC in $KC; do
-			if [ "$BBKC" = "cert" ]; then
-				FILEEXTENSION="crt";
-			else
-				FILEEXTENSION="key";
-			fi
-			if grep -q "^${BBKC}file" $CONF; then
-				sed -e "s/^${BBKC}file=.*/${BBKC}file=\/opt\/yaffas\/etc\/ssl\/certs\/$PRODUCT.${FILEEXTENSION}/" $CONF > $TMPCONF
-				rm -f $CONF
-				mv -f $TMPCONF $CONF
-				chmod 600 $CONF
-			else
-				echo "${BBKC}file=/opt/yaffas/etc/ssl/certs/$PRODUCT.${FILEEXTENSION}" >> $CONF
-			fi
-		done
-	fi
-done
+if [ $1 -eq 1 ]; then
+	set -e
+	CERTDIR="/opt/yaffas/etc/ssl/certs/"
+	CERTS="postfix exim webmin cyrus usermin ldap zarafa-webaccess zarafa-server zarafa-gateway zarafa-ical mppserver"
+	for i in $CERTS; do
+		ln -sf $CERTDIR/org/default.key $CERTDIR/$i.key
+		ln -sf $CERTDIR/org/default.crt $CERTDIR/$i.crt
+		cat $CERTDIR/org/default.crt $CERTDIR/org/default.key > $CERTDIR/$i
+	done
+	PRODUCTS="webmin usermin"
+	for PRODUCT in $PRODUCTS; do
+		CONF="/opt/yaffas/etc/$PRODUCT/miniserv.conf"
+		TMPCONF="/tmp/miniserv.conf"
+		if [ -r $CONF ]; then
+			echo "Editing $PRODUCT"
+			KC="key cert"
+			for BBKC in $KC; do
+				if [ "$BBKC" = "cert" ]; then
+					FILEEXTENSION="crt";
+				else
+					FILEEXTENSION="key";
+				fi
+				if grep -q "^${BBKC}file" $CONF; then
+					sed -e "s/^${BBKC}file=.*/${BBKC}file=\/opt\/yaffas\/etc\/ssl\/certs\/$PRODUCT.${FILEEXTENSION}/" $CONF > $TMPCONF
+					rm -f $CONF
+					mv -f $TMPCONF $CONF
+					chmod 600 $CONF
+				else
+					echo "${BBKC}file=/opt/yaffas/etc/ssl/certs/$PRODUCT.${FILEEXTENSION}" >> $CONF
+				fi
+			done
+		fi
+	done
+fi
 
 %postun
 if [ $1 -eq 0 ]; then
